Refactored the project to match other TS projects: - Set up editorconfig, linting, prettier rules - Renamed folders to match other projects - Changed unit testing stack from mocha/chai to jest - Added github actions for npm publishing and CI - Fixed linting errors - Other various minor changes
