+++ .eslintignore
@@ -1,0 +1,5 @@
+dist
+node_modules
+coverage
+.eslintrc.js
+generated
