+++ .eslintrc.js
@@ -1,0 +1,56 @@
+module.exports = {
+  root: true,
+  parser: "@typescript-eslint/parser",
+  parserOptions: {
+    tsconfigRootDir: __dirname,
+    project: ["./tsconfig.json"],
+  },
+  plugins: ["@typescript-eslint", "eslint-comments", "jest", "promise"],
+  extends: [
+    "airbnb-typescript/base",
+    "plugin:@typescript-eslint/recommended",
+    "plugin:eslint-comments/recommended",
+    "plugin:jest/recommended",
+    "plugin:promise/recommended",
+    "prettier",
+    "plugin:@typescript-eslint/recommended-requiring-type-checking",
+  ],
+  env: {
+    node: true,
+    browser: true,
+    jest: true,
+  },
+  rules: {
+    // note you must disable the base rule as it can report incorrect errors
+    "quotes": ["error", "double"],
+    "no-unused-vars": "off",
+    "@typescript-eslint/no-unused-vars": ["error"],
+    "no-underscore-dangle": "off",
+    "lines-between-class-members": ["warn", "always"],
+    "@typescript-eslint/explicit-function-return-type": ["error"],
+    "@typescript-eslint/naming-convention": [
+      "error",
+      { selector: "typeLike", format: ["PascalCase"] },
+      { selector: "enumMember", format: ["PascalCase"] },
+      { selector: "variableLike", format: ["camelCase"] },
+      {
+        selector: "interface",
+        format: ["PascalCase"],
+        custom: {
+          regex: "^I[A-Z]",
+          match: false,
+        },
+      },
+      {
+        selector: "typeAlias",
+        format: ["PascalCase"],
+        custom: {
+          regex: "^I[A-Z]",
+          match: false,
+        },
+      },
+    ],
+    "class-methods-use-this": "off",
+    "jest/no-conditional-expect": "off"
+  },
+};
