+++ .github/CODEOWNERS
@@ -1,0 +1,2 @@
+# Code owners:
+* @greenemcg @AndrewBird81 @adongare @jkotanchik-SB @serhii-ilin @RohitKandimalla
