+++ .gitignore
@@ -52,9 +52,6 @@ typings/
 
 node_modules
 
-lib
+dist
 .idea
-/tests/*.test.js
-/tests/*.test.js.map
-/tests/testCql.js.map
-/tests/testCql.js
+coverage
