+++ jest.config.js
@@ -1,0 +1,11 @@
+module.exports = {
+  preset: "ts-jest",
+  testEnvironment: "node",
+  roots: ["<rootDir>/src/", "<rootDir>/test/"],
+  modulePathIgnorePatterns: ["<rootDir>/dist/"],
+  collectCoverageFrom: [
+    "src/**/*.ts",
+    "!**/node_modules/**",
+  ],
+  setupFilesAfterEnv: ["jest-extended"],
+};
