+++ package.json
@@ -3,15 +3,17 @@
   "version": "0.0.1-alpha",
   "description": "Antlr Parsing of CQL in typescript",
   "files": [
-    "lib"
+    "dist"
   ],
   "scripts": {
-    "test": "mocha -r ts-node/register tests/**/*.test.ts",
-    "build-lib": "npx tsc -p tsconfig.json",
-    "coverage": "nyc -r text -e .ts -x \"tests/*.test.ts\" npm run test",
-    "antlr4ts": "antlr4ts src/antlr/generated/cql.g4"
+    "test": "npx jest",
+    "coverage": "npx jest --coverage",
+    "prebuild": "rm -rf dist",
+    "build": "tsc",
+    "generate": "npx antlr4ts generated/cql.g4",
+    "lint": "npx eslint . --fix --ext .js,.ts"
   },
-  "main": "lib/index.js",
+  "main": "dist/index.js",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/MeasureAuthoringTool/cql-antlr-parser.git"
@@ -21,22 +23,32 @@
     "CQL",
     "Fhir"
   ],
-  "author": "Michael Greene",
+  "author": "SemanticBits sb-mat-help@semanticbits.com",
   "license": "CREATIVE COMMONS ATTRIBUTION 4.0 INTERNATIONAL LICENSE",
   "bugs": {
     "url": "https://github.com/MeasureAuthoringTool/cql-antlr-parser/issues"
   },
   "homepage": "https://github.com/MeasureAuthoringTool/cql-antlr-parser#readme",
   "devDependencies": {
-    "@types/chai": "^4.2.15",
-    "@types/mocha": "^8.2.2",
-    "chai": "^4.3.4",
-    "mocha": "^8.3.2",
-    "nyc": "^15.1.0",
+    "@types/jest": "^26.0.22",
+    "@typescript-eslint/eslint-plugin": "^4.20.0",
+    "@typescript-eslint/parser": "^4.20.0",
+    "eslint": "^7.23.0",
+    "eslint-config-airbnb-typescript": "^12.3.1",
+    "eslint-config-prettier": "^8.1.0",
+    "eslint-plugin-eslint-comments": "^3.2.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-jest": "^24.3.2",
+    "eslint-plugin-promise": "^4.3.1",
+    "jest": "^26.6.3",
+    "jest-extended": "^0.11.5",
+    "prettier": "^2.2.1",
+    "ts-jest": "^26.5.4",
     "ts-node": "^9.1.1",
     "typescript": "^4.2.3"
   },
   "dependencies": {
-    "antlr4ts": "^0.5.0-alpha.4"
+    "antlr4ts": "^0.5.0-alpha.4",
+    "antlr4ts-cli": "^0.5.0-alpha.4"
   }
 }
