--- src/antlr/dao/CqlText.ts
@@ -1,9 +1,0 @@
-import LineInfo from "./LineInfo";
-
-export default interface CqlText {
-    text?: string;
-    name?: string;
-
-    start?: LineInfo;
-    stop?: LineInfo;
-}
