--- src/antlr/index.ts
@@ -1,12 +1,0 @@
-import CqlAntlr from "./CqlAntlr";
-import CqlCode from "./dao/CqlCode";
-import CqlCodeSystem from "./dao/CqlCodeSystem";
-import CqlInclude from "./dao/CqlInclude";
-import CqlResult from "./dao/CqlResult";
-import CqlText from "./dao/CqlText";
-import CqlValueSet from "./dao/CqlValueSet";
-import CqlVersion from "./dao/CqlVersion";
-import LineInfo from "./dao/LineInfo";
-
-export {CqlAntlr, CqlCode, CqlCodeSystem, CqlInclude, CqlResult, CqlText, CqlValueSet, CqlVersion, LineInfo};
-
