+++ src/dto/CqlCode.ts
@@ -1,6 +1,6 @@
 import CqlText from "./CqlText";
 
 export default interface CqlCode extends CqlText {
-    codeId?: string;
-    codeSystem?: string;
+  codeId?: string;
+  codeSystem?: string;
 }
