+++ src/dto/CqlCodeSystem.ts
@@ -1,5 +1,5 @@
 import CqlVersion from "./CqlVersion";
 
 export default interface CqlCodeSystem extends CqlVersion {
-    oid?: string;
+  oid?: string;
 }
