+++ src/dto/CqlInclude.ts
@@ -1,5 +1,5 @@
 import CqlVersion from "./CqlVersion";
 
 export default interface CqlInclude extends CqlVersion {
-    called?: string;
+  called?: string;
 }
