+++ src/dto/CqlText.ts
@@ -1,0 +1,8 @@
+import LineInfo from "./LineInfo";
+
+export default interface CqlText {
+  text?: string;
+  name?: string;
+  start?: LineInfo;
+  stop?: LineInfo;
+}
