+++ src/dto/CqlValueSet.ts
@@ -1,5 +1,5 @@
 import CqlText from "./CqlText";
 
 export default interface CqlValueSet extends CqlText {
-    url?: string;
+  url?: string;
 }
