+++ src/dto/CqlVersion.ts
@@ -1,5 +1,5 @@
 import CqlText from "./CqlText";
 
 export default interface CqlVersion extends CqlText {
-    version?: string;
+  version?: string;
 }
