+++ src/dto/LineInfo.ts
@@ -1,5 +1,4 @@
 export default interface LineInfo {
- line: number;
- position: number;
+  line: number;
+  position: number;
 }
-
